test(ocpp): split K28 tests into own TU
Move the 47 K28 dynamic-profile cases out of test_smart_charging.cpp
(2753->847 lines) into test_smart_charging_dynamic.cpp. Hoist the
shared SmartChargingTestV21 fixture into smart_charging_test_utils.hpp
so both TUs include it. Pure move; all 78 cases unchanged.
